+++ json_annotation/lib/src/allowed_keys_helpers.dart
@@ -3,7 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 /// Helper function used in generated code when
-/// `JsonAnnotation.disallowUnregognizedKeys` is `true`.
+/// `JsonAnnotation.disallowUnrecognizedKeys` is `true`.
 ///
 /// Should not be used directly.
 void $checkAllowedKeys(Map map, Iterable<String> allowedKeys) {